+++ psyneulink/core/components/functions/__init__.py
@@ -2,13 +2,19 @@ from . import function
 from . import userdefinedfunction
 from . import combinationfunctions
 from . import interfacefunctions
+from . import transferfunctions
+from . import selectionfunctions
 
 from .function import *
 from .userdefinedfunction import *
 from .combinationfunctions import *
 from .interfacefunctions import *
+from .transferfunctions import *
+from .selectionfunctions import *
 
 __all__ = list(function.__all__)
 __all__.extend(userdefinedfunction.__all__)
 __all__.extend(combinationfunctions.__all__)
 __all__.extend(interfacefunctions.__all__)
+__all__.extend(transferfunctions.__all__)
+__all__.extend(selectionfunctions.__all__)
