+++ psyneulink/core/components/functions/function.py
@@ -1181,11 +1181,6 @@ class ArgumentTherapy(Function_Base):
 # endregion
 
 
-# *******************************************  SELECTION FUNCTIONS *****************************************************
-
-
-# endregion ************************************************************************************************************
-
 # region ***********************************  INTEGRATOR FUNCTIONS *****************************************************
 
 #  Integrator
