+++ psyneulink/library/components/mechanisms/processing/transfer/kohonenmechanism.py
@@ -61,7 +61,8 @@ import numpy as np
 import typecheck as tc
 
 from psyneulink.core.components.functions.function import Kohonen, is_function_type
-from psyneulink.core.components.functions.transferfunctions import Linear, OneHot
+from psyneulink.core.components.functions.transferfunctions import Linear
+from psyneulink.core.components.functions.selectionfunctions import OneHot
 from psyneulink.core.components.mechanisms.adaptive.learning.learningmechanism import ACTIVATION_INPUT, ACTIVATION_OUTPUT, LearningMechanism
 from psyneulink.core.components.mechanisms.mechanism import Mechanism
 from psyneulink.core.components.mechanisms.processing.transfermechanism import TransferMechanism
